# tests: speedup persist huge tables test