--- a/spec/unit/persist_spec.lua
+++ b/spec/unit/persist_spec.lua
@@ -95,7 +95,7 @@ describe("Persist module", function()
     end)
 
     it("should work with huge tables", function()
-        local tab = arrayOf(100000)
+        local tab = arrayOf(10000)
         for _, codec in ipairs({"bitser", "luajit"}) do
             local ser = Persist.getCodec(codec).serialize
             local deser = Persist.getCodec(codec).deserialize