For example on my machine:

| test duration (ms) |  master  |             PR |
| :-                 | -:       | -:             |
| regular run        |   859.82 |  104.38 (÷8.2) |
| coverage run       | 40735.60 | 4151.34 (÷9.8) |